requirements.txt
selenium==4.37.0
bs4==0.0.2
requests==2.32.5
lxml==6.0.2
python-dotenv==1.2.1
undetected-chromedriver==3.5.5
driver==0.1.0
setuptools==80.9.0